encoding: Move ChannelMask from maccommands to types module
