+++ lorawan-encoding/src/parser.rs
@@ -21,7 +21,8 @@
 //! ```
 
 use super::keys::{AppKey, AppSKey, CryptoFactory, Encrypter, NwkSKey, AES128, MIC};
-use super::maccommands::{ChannelMask, DLSettings, Frequency};
+use super::maccommands::{DLSettings, Frequency};
+use crate::types::ChannelMask;
 
 use super::securityhelpers;
 use super::securityhelpers::generic_array::GenericArray;
